refactor(shared): 将randomId方法抽离到shared/utils目录中
