+++ packages/devui-vue/devui/shared/util/index.ts
@@ -1,0 +1,8 @@
+export * from './class';
+export * from './common-var';
+export * from './debounce';
+export * from './dom';
+export * from './lock-scroll';
+export * from './props-util';
+export * from './random-id';
+export * from './set-style';
