+++ packages/devui-vue/devui/tree/src/core/utils.ts
@@ -1,4 +1,4 @@
-import { randomId } from '../../../anchor/src/util';
+import { randomId } from '../../../shared/util';
 import { IInnerTreeNode, ITreeNode } from './use-tree-types';
 
 export function flatToNested(flatTree: IInnerTreeNode[]): ITreeNode[] {
